Remove `core.allowWindowTransparency` from the config schema… (#1206)
…so that it's not listed among the core settings in `settings-view`.

We'll leave the feature in, but users will have to opt into it by editing their `config.cson` or setting it via the dev tools console.

This is a half-baked setting that leads to bad experiences for new Pulsar users. It can stay in for power users who've figured out how to mitigate the bad side effects.