CI: Delete hardcoded GITHUB_TOKEN reference (Cirrus)
We can define this on the web dashboard instead to save ourselves the
trouble of updating it in the file every time it expires.

This token isn't particularly sensitive, as it has no permissions.
It's only for authing as a real user so rate limits are relaxed,
letting us download vscode-ripgrep from an otherwise *heavily*
rate-limited datacenter / "cloud" machine that we call "CI."

For the more sensitive tokens, I'd like to be able to apply them
per-step. Which I don't _think_ the Cirrus web dashboard lets you do.
This one should be alright to define globally.
(Heck, we already do define this one for the whole CI build file!
Moving it to the web dashboard won't change that.)