CI: Compile and install newer Python for ARM Linux Cirrus
Need this for compatibility with recent gyp-next, from the node-gyp
bump in the latest ppm bump for Pulsar core.

That dependency bump improved our out-of-the-box support for
Python 3.12+, relevant when building packages with native C/C++
addon modules while installing said packages with ppm, or the GUI
package management interface in Pulsar settings (which also uses ppm
under the hood anyway).

We will cache the compiled Python in Cirrus's cache storage system,
 much like we cache compiled Python in GitHub Actions caches.
Want to save time and datacenter power use/emissions and such.

Known SHA256 hash of the Python source tarball is hardcoded and checked
for safety/reliability. We can update that hash if we need to build
a different Python version in the future.