CI: Install Python setuptools on ARM Linux Cirrus
We are using old Node in CI, which comes with old npm, which comes with
old node-gyp. Yarn uses npm's (in this case, old) copy of node-gyp
by default, so we still need to have Python setuptools installed
for compatibility with old node-gyp x new Python (3.12+) in CI.

Could do 'yarn global add node-gyp@10.2.0' instead, but this also works.