Purge a stale fold boundaries iterator…
…instead of silently failing.

For some reason, a fold cache wasn't updating in PHP in certain circumstances. We were trying to resolve a fold on a given row by inspecting a capture whose node was useless to inspect because its tree had long since been invalidated.

This was only reproducible on startup, so maybe it has to do with non-root language layers not having their fold caches purged after tokenization?

This isn't easy to reproduce in a way that allows us to write a spec, so I've written this code very conservatively so that it can't possibly regress — only do slightly more work than it used to.

One way around this would be to extract the metadata we need when we create the boundaries iterator (so that we don't have to inspect the node later). That would theoretically increase the lifetime of the cache, but would involve doing more work up front. So it may or may not be worth it performance-wise. (It's not like the folds query is particularly costly, and it might be pointless to try to extend the cache's lifespan when code changes constantly and the _underlying data_ will be stale after a short while.)