Tree-sitter rolling fixes, 1.127 edition (#1240)
* Upgrade to `web-tree-sitter` v0.25.3

This is a big upgrade! `web-tree-sitter` has been rewritten in TypeScript and has made some backwards-incompatible changes to its exports — all of which are good things.

I've also updated the instructions for generating the `web-tree-sitter` bundle.

* [language-python] Various folds fixes…

* Make function parameters spanning multiple lines fold properly
* Make `try`/`except` clauses fold properly, with `try` folding only up to the first `except`
* Add `folding-spec.js` to test these cases and others and guard them against regressions

Also fix a silly mistake in an indentation query.

Fixes #1216.

* [language-python] Convert specs to Jasmine 2 runner

* Get grammar specs working on Jasmine 2 runner

* Monkey-patch `Query::captures` more directly

* [language-ruby] Properly highlight and indent new `case/in` syntax

* [language-ruby] Properly fold `case` statements and `when/if` clauses

* Add logging to compare/contrast CI output between Pulsar and PulsarNext

* [language-javascript] Ensure `foo` in `this.#foo = 0` highlights properly

* Purge a stale fold boundaries iterator…

…instead of silently failing.

For some reason, a fold cache wasn't updating in PHP in certain circumstances. We were trying to resolve a fold on a given row by inspecting a capture whose node was useless to inspect because its tree had long since been invalidated.

This was only reproducible on startup, so maybe it has to do with non-root language layers not having their fold caches purged after tokenization?

This isn't easy to reproduce in a way that allows us to write a spec, so I've written this code very conservatively so that it can't possibly regress — only do slightly more work than it used to.

One way around this would be to extract the metadata we need when we create the boundaries iterator (so that we don't have to inspect the node later). That would theoretically increase the lifetime of the cache, but would involve doing more work up front. So it may or may not be worth it performance-wise. (It's not like the folds query is particularly costly, and it might be pointless to try to extend the cache's lifespan when code changes constantly and the _underlying data_ will be stale after a short while.)

* [language-php] Fold multiline comments

* Remove temporary logging