[language-json] Add optional comment support…
…plus an extra “JSON With Comments” grammar that explicitly allows them and is used for `.jsonc` files.

Comments are not allowed in the JSON spec, so not all tooling supports parsing JSON wiht comments. But that limits their usefulness as configuration files. VS Code has popularized JSONC (a flavor of JSON where both line and block comments are expressly permitted) as a config file format.

Currently, we do _nothing_ around comments in JSON files. They are highlighted like plain text. But `tree-sitter-json` tolerates and parses comments!

Because of this, we can support JSONC for free. Hence a new grammar that uses the exact same parser and query files as the current JSON grammar, but where comments are legal.

We can also handle them better in ordinary JSON files: by marking them with the `invalid.illegal` scope, much like we already do with trailing commas.

Some tooling even blurs lines a bit by adding comments to `.json` files. (`tsconfig.json`, from a _different_ Microsoft team, is perhaps the most notable offender.) So we’re also offering a way for users to opt into a mode that tolerates comments. A new “Allow Comments in .json files” setting on the package will stop marking comments as illegal in JSON.

I think this strikes a good balance between being correct and being realistic to the real world.