[language-(java|type)script] Fix overeager “invalid” highlighting…
…when Tree-sitter pretends to see an “optional chain” node in the tree.

Tree-sitter has to make the best of invalid syntax all the time because it reparses even while people are typing. When there are invalidities, it tries a number of strategies to get back to a normal parsing mode as quickly as possible.

The strategies fall into two buckets: (a) pretending that an invalid token isn't there, and (b) pretending something that _ought_ to be there actually _is_. If the latter strategy gets it to a valid tree faster, that's what it'll do.

`tree-sitter-javascript` parses `optional_chain` nodes (the “?.” in `foo?.bar`) in more places than they’re actually valid. They’re not allowed on the left-hand side of an assignment, so when we see one in that context, we add an `invalid.illegal` scope.

But sometimes the `optional_chain` node we see is not actually there because of Tree-sitter’s error-recovery behavior. So we add an `invalid.illegal` scope at a given position, expecting that we'll close it later… but we never do, since the closing scope boundary is at the same spot, and we close scopes before we open them.

There's a larger bug fix that needs to happen to address all _that_, but the simplest fix for now is to enforce the content of the `optional_chain` node so that we don’t try to add a scope to an empty node.