Don't store any scope boundaries if the capture range is empty
This is the “proper” fix for the issue I described in an earlier commit in which a zero-width capture could “leak” a scope name into everything that comes after it.

Those captures may still be valid for other uses. For instance, indentation resolution uses `ScopeResolver#store` to apply the logic of custom predicates, but does not need to consume the boundary data; it is happy to tolerate empty captures if the grammar author opts into it. That should be unchanged; we just won't add any boundary data that we know will produce invalidity.

This will do the right thing with regard to the syntax highlighting _and_ the “Editor: Log Cursor Scope” command.