Don't test for marker validity…
…when gathering existing injection markers to re-evaluate.

This was silly of me. When we repopulate injections, we cast as wide a net as possible to make sure we've caught all the existing child `LanguageLayer`s in the area affected by the buffer change. We do this for two reasons:

1. Often we can skip creating a new `LanguageLayer` if the old one is still valid and covers the range we expect.
2. The ones that _don't_ get matched up and reused need to be destroyed!

I wasn't thinking about the second part. I assumed that `MarkerLayer#findMarkers` wouldn't return invalid markers, but it does; and it's a good thing, too, because markers can go from invalid to valid (e.g., when the user navigates backward through the history stack).

Prematurely winnowing the invalid markers out of the set meant that some `LanguageLayer`s did not get destroyed when they should’ve been, leading to some redundant highlighting on inline Markdown code as I was writing it.

I tracked it down by re-introducing the old logic as a sort of heckler from the wings; whenever the new logic produced results inconsistent with what was there before, I had it log to the console and call me a fraud. It worked!