CI: Compile newer Python for Debian Docker image
(manual port of #1223)

Newer node-gyp (as a subdepenency of ppm) now supports Python 3.12!
But drops support for Python 3.7, which Debian 10 "Buster" uses.

Compile and install newer (not EOL, and supported by node-gyp)
Python in the Debian Docker image we build the Linux bins on.

(It's a bit slow and compute-intensive. The kind of thing best cached,
so we don't have to wait and waste a bunch of watts/heat/emissions
at some datacenter somewhere.)

So we will do it once and cache the result in GitHub Actions cache.

We can check SHA256 sum of Python tarball before using it, for safety.

We always install Python compilation dependencies, as apparently
we need these to do `make install` too, not just the main `make`.