Fix issue with Atom IDE usage of `HighlightIterator`…
…by providing a fallback value for the second argument of `seek`.

Atom IDE has a very creative strategy for doing headless imperative syntax highlighting: it creates a `TextBuffer`, sets the correct grammar and language mode, waits for the contents to parse, then calls `buildHighlightIterator` and seeks manually from point A to point B. You can see this in action in `atom-ide-datatip` and `atom-ide-signature-help`.

But this is an old strategy that assumes we're always working with `TextMateHighlightIterator`s — which need to be told where to start but not where to end.

The `HighlightIterator` used in `WASMTreeSitterLanguageMode` has a second parameter to `seek` that specifies where our highlighting job will end. If this parameter is omitted, syntax highlighting fails.

So we'll recover in these cases by setting our own end row. All of our own usages of `HighlightIterator::seek` supply two arguments, so this is just about supporting a very unusual strategy used by old code.