Prefer `@electron/remote`, deprecate direct usage of `electron.remote`…
…and switch all internal usages to `@electron/remote`.

This commit also includes the entire bundle of changes to the `tree-view` package — the decaffeination of the `tree-view` specs made it very hard to isolate this change atomically, so I'm reluctantly deciding to bring the whole thing along.

`tree-view` has been updated to fix a few bugs with Electron 30 (moving items to the trash, revealing items in the file explorer, etc.) and the specs have been fully decaffeinated (not just machine-decaffeinated).