Point to Electron `30.0.9`…
…and point our dependencies to the NPM-published versions under the `@pulsar-edit` namespace that are compatible with Electron 30.0.9.

This vastly reduces the number of random GitHub commits we're pointing to in our `dependencies` and helps the `resolutions` section be as streamlined as it can be.

Also includes `@electron/remote` and `@electron/rebuild`.

All internal references to paths are also changed in this commit — e.g., `require('foo')` becomes `require('@pulsar-edit/foo')`.