Eliminate reliance on `typescript-simple`…
…in favor of a custom solution that uses the `typescript` package directly.

`typescript-simple` is no longer updated and is arguably unnecessary anyway; it wasn't hard to replicate its behavior through direct consumption of `typescript`.

This preserves our ability to transpile TypeScript just like we can transpile CoffeeScript or `@babel`-ified JS.

This is not the happy path of TypeScript transpilation (do it yourself as a build step!) but it's behavior that exists, so we should ensure it continues to work.