Consume `get-scrollbar-style`…
…to replace the deprecated `scrollbar-style` module.

macOS offers two different “styles” of scrollbar — traditional scrollbars and “overlay” scrollbars that do not use a dedicated scrollbar track. Strangely, we can detect when the user switches from one to the other without any external dependencies… but to find out imperatively which one is currently active, we need a very light module with native bindings.

This commit adopts the new approach for detecting which macOS scrollbar style is enabled. Compared to the _status quo ante_, this should carry no behavior change on either macOS or other platforms.