Update to Electron 30 (#1367)
* Remove all support for legacy Tree-sitter

Abandoning support for the legacy Tree-sitter system  is a necessary step toward upgrading past Electron 12.2.3.

* Prefer `@electron/remote`, deprecate direct usage of `electron.remote`…

…and switch all internal usages to `@electron/remote`.

This commit also includes the entire bundle of changes to the `tree-view` package — the decaffeination of the `tree-view` specs made it very hard to isolate this change atomically, so I'm reluctantly deciding to bring the whole thing along.

`tree-view` has been updated to fix a few bugs with Electron 30 (moving items to the trash, revealing items in the file explorer, etc.) and the specs have been fully decaffeinated (not just machine-decaffeinated).

* Fix a failing `autocomplete-plus` spec

* Get `dev-live-reload` specs passing

* Get `image-view` specs passing

* [language-ruby] Add a basic grammar test spec

* [markdown-preview] Get specs passing

* [settings-view] Get specs passing

* Point to Electron `30.0.9`…

…and point our dependencies to the NPM-published versions under the `@pulsar-edit` namespace that are compatible with Electron 30.0.9.

This vastly reduces the number of random GitHub commits we're pointing to in our `dependencies` and helps the `resolutions` section be as streamlined as it can be.

Also includes `@electron/remote` and `@electron/rebuild`.

All internal references to paths are also changed in this commit — e.g., `require('foo')` becomes `require('@pulsar-edit/foo')`.

* Fix some `TextEditorComponent` issues…

…related to changes in Chromium sometime between Electron 12 and Electron 30.

For instance, text bounding-box logic changed at some point to require that our four “sentinel” characters be isolated from one another — otherwise character height/width measurements are inaccurate.

Also fixed some issues related to device pixel ratios that resulted in spec failures.

* Eliminate reliance on `typescript-simple`…

…in favor of a custom solution that uses the `typescript` package directly.

`typescript-simple` is no longer updated and is arguably unnecessary anyway; it wasn't hard to replicate its behavior through direct consumption of `typescript`.

This preserves our ability to transpile TypeScript just like we can transpile CoffeeScript or `@babel`-ified JS.

This is not the happy path of TypeScript transpilation (do it yourself as a build step!) but it's behavior that exists, so we should ensure it continues to work.

* Consume `get-scrollbar-style`…

…to replace the deprecated `scrollbar-style` module.

macOS offers two different “styles” of scrollbar — traditional scrollbars and “overlay” scrollbars that do not use a dedicated scrollbar track. Strangely, we can detect when the user switches from one to the other without any external dependencies… but to find out imperatively which one is currently active, we need a very light module with native bindings.

This commit adopts the new approach for detecting which macOS scrollbar style is enabled. Compared to the _status quo ante_, this should carry no behavior change on either macOS or other platforms.

* Additional changes related to device pixel ratio…

…that probably should have gone in the earlier commit.

* Add the ability to report crashes…

…via Electron's standard `crashReporter` interface.

Crash reports are enabled on Windows and Linux by default (macOS’s own crash reports via `Console.app` are typically more helpful, but go away if crash reporting is enabled). But you can opt into crash reporting explicitly on all platforms via the `--crashdump` command-line switch.

* Delete deprecated "app.allowRendererProcessReuse = false"

This option has been removed, and thus no-longer does anything,
as of Electron 14.

Committer note: This change was split out from the prior commit.

* Enable `@electron-remote` for the given window during startup…

…and disable a Blink experimental feature that was causing trouble with package compatibility (because it added an `on` method to all DOM nodes!).

* Be more paranoid about release channels in a handful of places…

…in case we want to maintain a `next` (or `nightly` or `preview`) release channel on an ongoing basis.

* Cache `Language` instances in `web-tree-sitter`…

…so that the specs don't run out of memory.

Also be more diligent about cleaning up parsers that we created for a given buffer when that buffer is destroyed.

* Preserve an `ATOM_CHANNEL` environment variable if it exists…

…and other changes related to release-channel agnosticism.

(These should've been included in an earlier commit.)

* [snippets] Get specs passing

* [spell-check] Get specs passing…

…and abandon `atom-pathspec` in favor of an internal implementation.

(We are pinning to a specific commit of `node-spellchecker`. This one escaped my notice during the context-aware-native-modules sweep — probably because it's in a package! We should revisit this and publish it to NPM instead, but this is likely good enough for now.)

* [tabs] Fix an exception that can happen in unusual scenarios…

…when the browser can't give us an element under a given X/Y coordinate.

* [update-package-dependencies] Get specs passing

* [symbol-provider-tree-sitter] Unmock clock in the standard way

* Remove outdated Tree-sitter documentation

* Update GitHub Actions CI jobs…

…to include Wayland-related keyboard dependencies.

(This probably should be done for the Cirrus builds as well!)

* Update CirrusCI build job…

…to include Wayland-related keyboard dependencies.

(Untested, but copied directly from the GHA workflow files.)

* Revert to a single `entitlements.plist`…

…now that we're no longer affected by the awful `libuv` bug that was present in Electron 12.

* Update `script/electron-builder.js`…

…to support a couple of command-line switches and to build correctly for Electron 30.

* Fix `ThemeManager` specs

* Define two methods on the `atom` global…

…to work around `tree-view` issues related to trashing and revealing items in `electron.shell`.

`atom.trashItem` and `atom.showItemInFolder` are drop-in replacements for `electron.shell.trashItem` and `electron.shell.showItemInFolder`.

TODO: These might not be necessary; try the `@electron/remote` versions instead.

* Comment reformatting

* Some integration test fixes

* Fix `PackageManager` specs

* Use `yarn.lock` from `updated-latest-electron` branch

* Swallow exceptions raised by pending tests

* Disallow `if(` and the like in `.eslitrc`

* Add `submodules: true` back into certain workflows

* Remove `atom.trashItem` and `atom.showItemInFolder`…

…since they weren't even being used. The equivalent functions from the `shell` module of `@electron/remote` have been performing the same task in PulsarNext for a while now.

* Update ppm to commit d288b0978b35f7c0ffb3dc5c1e50c

Includes ppm PRs:
- 154 "Fix CI"
- 155 "Fix misnamed variable in `rebuild` command"

* Update ppm to commit 07faba10096ee4239ce671218d9953fb2e40ecbf

* Remove unnecessary `Promise.all`

---------

Co-authored-by: Maurício Szabo <mauricioszabo@users.noreply.github.com>
Co-authored-by: DeeDeeG <DeeDeeG@users.noreply.github.com>